Relocate Guzhenren abilities into GuScript module
